// apps/admin/src/api/generated/schemas/planListResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { PlanListResponseData } from './planListResponseData';
import type { PlanListResponseMessage } from './planListResponseMessage';
import type { PlanListResponseError } from './planListResponseError';
import type { PlanListResponseMeta } from './planListResponseMeta';
import type { PlanListResponseRequestId } from './planListResponseRequestId';

export interface PlanListResponse {
  success?: boolean;
  data?: PlanListResponseData;
  message?: PlanListResponseMessage;
  error?: PlanListResponseError;
  meta?: PlanListResponseMeta;
  requestId?: PlanListResponseRequestId;
  timestamp?: string;
}