// apps/admin/src/api/generated/schemas/settingItem.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Auto-generated from Flask backend routes
 * OpenAPI spec version: 1.0.0
 */
import type { SettingItemCategory } from './settingItemCategory';
import type { SettingItemIsPublic } from './settingItemIsPublic';

export interface SettingItem {
  key: string;
  value: string;
  category?: SettingItemCategory;
  isPublic?: SettingItemIsPublic;
}