// apps/admin/src/api/generated/schemas/systemSettingRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { SystemSettingReadCreatedAt } from './systemSettingReadCreatedAt';
import type { SystemSettingReadUpdatedAt } from './systemSettingReadUpdatedAt';
import type { SystemSettingReadValue } from './systemSettingReadValue';
import type { SystemSettingReadDescription } from './systemSettingReadDescription';
import type { SystemSettingReadCategory } from './systemSettingReadCategory';

/**
 * Schema for reading SystemSetting
 */
export interface SystemSettingRead {
  /** Creation timestamp */
  createdAt?: SystemSettingReadCreatedAt;
  /** Last update timestamp */
  updatedAt?: SystemSettingReadUpdatedAt;
  key: string;
  value?: SystemSettingReadValue;
  description?: SystemSettingReadDescription;
  category?: SystemSettingReadCategory;
  isPublic?: boolean;
}