// apps/admin/src/api/generated/schemas/responseEnvelopeListSystemSettingRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeListSystemSettingReadData } from './responseEnvelopeListSystemSettingReadData';
import type { ResponseEnvelopeListSystemSettingReadMessage } from './responseEnvelopeListSystemSettingReadMessage';
import type { ResponseEnvelopeListSystemSettingReadError } from './responseEnvelopeListSystemSettingReadError';
import type { ResponseEnvelopeListSystemSettingReadMeta } from './responseEnvelopeListSystemSettingReadMeta';
import type { ResponseEnvelopeListSystemSettingReadRequestId } from './responseEnvelopeListSystemSettingReadRequestId';

export interface ResponseEnvelopeListSystemSettingRead {
  success?: boolean;
  data?: ResponseEnvelopeListSystemSettingReadData;
  message?: ResponseEnvelopeListSystemSettingReadMessage;
  error?: ResponseEnvelopeListSystemSettingReadError;
  meta?: ResponseEnvelopeListSystemSettingReadMeta;
  requestId?: ResponseEnvelopeListSystemSettingReadRequestId;
  timestamp?: string;
}